+++ jubeatools/formats/konami/eve/tests/test_eve.py
@@ -2,7 +2,7 @@ from hypothesis import given
 
 from jubeatools import song
 from jubeatools.formats import Format
-from jubeatools.formats.konami.testutils import eve_compatible_song, open_temp_dir
+from jubeatools.formats.konami.testutils import eve_compatible_song
 from jubeatools.testutils.test_patterns import dump_and_load_then_compare
 
 
@@ -11,7 +11,6 @@ def test_that_full_chart_roundtrips(song: song.Song) -> None:
     dump_and_load_then_compare(
         Format.EVE,
         song,
-        temp_path=open_temp_dir(),
         bytes_decoder=lambda b: b.decode("ascii"),
         load_options={"beat_snap": 12},
     )